refactor CurrencyWithSmallCents.jsx and parameterize
