+++ frontend/src/components/PortfolioFundingTotal/util.js
@@ -8,8 +8,3 @@ export const getCurrentFiscalYear = (today) => {
 
     return fiscalYear;
 };
-
-export const getCents = (amount) => {
-    // amount is a float currency value - returns the cents as an integer
-    return (amount - parseInt(amount)) * 100;
-};
