+++ frontend/src/components/UI/CurrencyWithSmallCents/util.js
@@ -1,0 +1,4 @@
+export const getCents = (amount) => {
+    // amount is a float currency value - returns the cents as an integer
+    return (amount - parseInt(amount)) * 100;
+};
